[CSL-1871] Reorganize slotting modules in cardano-sl-core
