+++ core/Pos/Core/Block/Main/Chain.hs
@@ -26,9 +26,10 @@ import           Pos.Core.Block.Union.Types (Block, BlockHeader, BlockSignature 
 import           Pos.Core.Class (IsMainHeader (..))
 import           Pos.Core.Configuration (HasConfiguration)
 import           Pos.Core.Delegation (DlgPayload)
+import           Pos.Core.Slotting.Types (SlotId (..))
 import           Pos.Core.Ssc (SscPayload, SscProof, mkSscProof)
 import           Pos.Core.Txp (TxPayload, TxProof, mkTxProof)
-import           Pos.Core.Types (ChainDifficulty, SlotId (..))
+import           Pos.Core.Types (ChainDifficulty)
 import           Pos.Core.Update (UpdatePayload, UpdateProof, mkUpdateProof)
 import           Pos.Crypto (Hash, PublicKey, hash)
