+++ core/Pos/Core/Class.hs
@@ -24,8 +24,8 @@ import           Universum
 
 import           Control.Lens (Getter, choosing, to)
 
-import           Pos.Core.Types (BlockVersion, ChainDifficulty, EpochIndex, EpochOrSlot (..),
-                                 HeaderHash, SlotId, SoftwareVersion)
+import           Pos.Core.Slotting.Types (EpochIndex, EpochOrSlot (..), SlotId)
+import           Pos.Core.Types (BlockVersion, ChainDifficulty, HeaderHash, SoftwareVersion)
 import           Pos.Crypto.Signing (PublicKey)
 import           Pos.Util.Some (Some, applySome, liftLensSome)
