+++ core/Pos/Core/Genesis/Types.hs
@@ -36,9 +36,10 @@ import           Serokell.Util (allDistinct, mapJson)
 import           Pos.Binary.Crypto ()
 import           Pos.Core.Coin ()
 import           Pos.Core.Delegation.Types (ProxySKHeavy)
+import           Pos.Core.Slotting.Types (Timestamp)
 import           Pos.Core.Ssc.Types (VssCertificatesMap, getVssCertificatesMap)
 import           Pos.Core.Types (Address, BlockVersionData, Coin, CoinPortion, SharedSeed,
-                                 StakeholderId, Timestamp)
+                                 StakeholderId)
 import           Pos.Crypto.Configuration (ProtocolMagic)
 import           Pos.Crypto.Signing (RedeemPublicKey)
