+++ core/Pos/Core/Ssc/Types.hs
@@ -52,7 +52,8 @@ import           Serokell.Util (listJson)
 
 import           Pos.Binary.Class (AsBinary (..), Bi, serialize')
 import           Pos.Core.Address (addressHash)
-import           Pos.Core.Types (EpochIndex, StakeholderId)
+import           Pos.Core.Slotting.Types (EpochIndex)
+import           Pos.Core.Types (StakeholderId)
 import           Pos.Crypto (DecShare, EncShare, Hash, PublicKey, Secret, SecretProof, Signature,
                              VssPublicKey, shortHashF)
