+++ infra/Pos/DHT/Workers.hs
@@ -17,8 +17,7 @@ import           Pos.Binary.Class (serialize)
 import           Pos.Binary.Infra.DHTModel ()
 import           Pos.Communication.Protocol (OutSpecs, WorkerSpec, localOnNewSlotWorker)
 import           Pos.Core.Configuration (HasConfiguration)
-import           Pos.Core.Slotting (flattenSlotId)
-import           Pos.Core.Types (slotIdF)
+import           Pos.Core.Slotting (flattenSlotId, slotIdF)
 import           Pos.DHT.Configuration (kademliaDumpInterval)
 import           Pos.DHT.Real.Types (KademliaDHTInstance (..))
 import           Pos.Infra.Configuration (HasInfraConfiguration)
