+++ infra/Pos/Slotting/Impl/Sum.hs
@@ -20,15 +20,14 @@ module Pos.Slotting.Impl.Sum
 
 import           Universum
 
-import           Ether.Internal (HasLens (..))
-
-import           Pos.Core.Types (SlotId (..), Timestamp)
+import           Pos.Core.Slotting (SlotId (..), Timestamp)
 import           Pos.Slotting.Impl.Ntp (NtpMode, NtpSlottingVar, NtpWorkerMode, ntpCurrentTime,
                                         ntpGetCurrentSlot, ntpGetCurrentSlotBlocking,
                                         ntpGetCurrentSlotInaccurate, ntpWorkers)
 import           Pos.Slotting.Impl.Simple (SimpleSlottingMode, SimpleSlottingVar,
                                            currentTimeSlottingSimple, getCurrentSlotBlockingSimple,
                                            getCurrentSlotInaccurateSimple, getCurrentSlotSimple)
+import           Pos.Util (HasLens (..))
 
 -- | Sum of all contexts used by slotting implementations.
 data SlottingContextSum
