+++ infra/Pos/Slotting/Impl/Util.hs
@@ -11,9 +11,9 @@ import           Data.Time.Units (Microsecond, convertUnit)
 import           NTP.Example ()
 
 import           Pos.Core.Configuration (HasConfiguration, epochSlots)
-import           Pos.Core.Slotting (flattenEpochIndex, mkLocalSlotIndex, unflattenSlotId)
-import           Pos.Core.Timestamp (addTimeDiffToTimestamp)
-import           Pos.Core.Types (EpochIndex, LocalSlotIndex, SlotId (..), Timestamp (..))
+import           Pos.Core.Slotting (EpochIndex, LocalSlotIndex, SlotId (..), Timestamp (..),
+                                    addTimeDiffToTimestamp, flattenEpochIndex, mkLocalSlotIndex,
+                                    unflattenSlotId)
 import           Pos.Util.Util (leftToPanic)
 
 import           Pos.Slotting.MemState (MonadSlotsData, getSystemStartM, withSlottingVarAtomM)
@@ -139,4 +139,3 @@ slotFromTimestamp approxCurTime = do
 
         epochDuration :: Microsecond
         epochDuration = slotDuration * fromIntegral epochSlots
-
