+++ lrc/Pos/Lrc/DB/Common.hs
@@ -29,7 +29,7 @@ import qualified Database.RocksDB as Rocks
 import           Pos.Binary.Class (Bi)
 import           Pos.Binary.Core ()
 import           Pos.Core.Configuration (HasConfiguration)
-import           Pos.Core.Types (EpochIndex)
+import           Pos.Core.Slotting (EpochIndex)
 import           Pos.DB (dbSerializeValue)
 import           Pos.DB.Class (DBTag (LrcDB), MonadDB (dbDelete, dbWriteBatch), MonadDBRead (dbGet))
 import           Pos.DB.Error (DBError (DBMalformed))
