+++ lrc/Pos/Lrc/DB/RichmenBase.hs
@@ -21,7 +21,7 @@ import           Universum
 
 import           Pos.Binary.Class (serialize')
 import           Pos.Binary.Core ()
-import           Pos.Core.Types (EpochIndex)
+import           Pos.Core.Slotting (EpochIndex)
 import           Pos.DB.Class (MonadDB, MonadDBRead)
 import           Pos.Lrc.Class (RichmenComponent (..))
 import           Pos.Lrc.DB.Common (getBi, putBi)
