+++ update/Pos/Update/Poll/PollState.hs
@@ -28,7 +28,8 @@ import           Control.Lens (makeLenses)
 import qualified Data.HashMap.Strict as HM
 import qualified Data.HashSet as HS
 
-import           Pos.Core.Types (ApplicationName, BlockVersion, BlockVersionData, EpochIndex,
+import           Pos.Core.Slotting (EpochIndex)
+import           Pos.Core.Types (ApplicationName, BlockVersion, BlockVersionData,
                                  NumSoftwareVersion, SoftwareVersion (..), StakeholderId)
 import           Pos.Core.Update (UpId, UpdateProposal (..))
 import           Pos.Lrc.DB.Issuers (IssuersStakes)
